requirements.txt
streamlit
openai
python-dotenv
langchain
langchain-community
sentence-transformers
faiss-cpu
PyMuPDF
docx2txt
